+++ src/apis/index.ts
@@ -1,2 +1,1 @@
 export * from './Order';
-export * from './Fcm';
